integrate tickStats into timeSeriesStats and revert some renamings
