+++ packages/framework/src/utils/stats/index.ts
@@ -1,4 +1,4 @@
 export * from './accountTimeSeries.js'
-export * from './timeFrameStats.js'
+export * from './timeSeriesStats.js'
 export * from './dal/index.js'
 export * from './types.js'
